requirements.txt
assertpy==1.1
pytest==7.2.0
requests==2.28.1
google-cloud-storage==2.6.0

